fix(shared-db): removal policy

+++ packages/cdk-shared-database/src/lib/cdk-shared-database.OnEventVpc.ts
@@ -16,7 +16,11 @@ export const handler: CdkCustomResourceHandler = async (event) => {
     case 'Update':
       return onCreate(props);
     case 'Delete':
-      return onDelete(props);
+      if (event.ResourceProperties.removalPolicy === 'destroy') {
+        return onDelete(props);
+      } else {
+        return noop(props);
+      }
   }
 };
 
@@ -39,6 +43,11 @@ async function onDelete(props: EventProps): Promise<CdkCustomResourceResponse> {
     `Engine '${props.instanceConnectionObject.engine}' is not implemented`
   );
   await provider.delete(props);
+  return {
+    PhysicalResourceId: props.instanceConnectionObject.username,
+  };
+}
+async function noop(props: EventProps): Promise<CdkCustomResourceResponse> {
   return {
     PhysicalResourceId: props.instanceConnectionObject.username,
   };
